Implemented changes for only owner only edit/delete images

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,13 +92,25 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession httpSession) {
         Image image = imageService.getImage(imageId);
 
+        //error message initialization
+        String error = "Only the owner of the image can edit the image";
+        User user = (User) httpSession.getAttribute("loggeduser");
         String tags = convertTagsToString(image.getTags());
+
+
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
-        return "images/edit";
+        //Adding error string to model object
+        model.addAttribute("editError", error);
+        //implementation of validation for authorize user only edit the image
+        if(user.getId().equals(image.getUser().getId())) {
+            return "images/edit";
+        } else {
+            return "images/image";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,9 +152,25 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession httpSession) {
+        //error message initialization
+        String error = "Only the owner of the image can delete the image";
+        User user = (User) httpSession.getAttribute("loggeduser");
+
+        Image image = imageService.getImage(imageId);
+
+        model.addAttribute("user", user);
+        model.addAttribute("image", image);
+        //Adding error string to model object
+        model.addAttribute("deleteError", error);
+
+        //implementation of validation for authorize user only delete the image
+        if (user.getId().equals(image.getUser().getId())) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images/";
+        } else {
+            return "images/image";
+        }
     }
